Move oauthtoken helpers to their own package
