+++ internal/authz/providers/bitbucketcloud/BUILD.bazel
@@ -13,6 +13,7 @@ go_library(
         "//internal/authz",
         "//internal/authz/types",
         "//internal/database",
+        "//internal/database/oauthtoken",
         "//internal/extsvc",
         "//internal/extsvc/auth",
         "//internal/extsvc/bitbucketcloud",
