+++ internal/authz/providers/github/BUILD.bazel
@@ -15,6 +15,7 @@ go_library(
         "//internal/authz",
         "//internal/authz/types",
         "//internal/database",
+        "//internal/database/oauthtoken",
         "//internal/encryption/keyring",
         "//internal/extsvc",
         "//internal/extsvc/auth",
