+++ internal/database/BUILD.bazel
@@ -46,7 +46,6 @@ go_library(
         "mockerr.go",
         "namespace_permissions.go",
         "namespaces.go",
-        "oauth_token_helper.go",
         "org_invitations.go",
         "org_members.go",
         "orgs.go",
@@ -129,10 +128,8 @@ go_library(
         "//internal/featureflag",
         "//internal/github_apps/store",
         "//internal/hashutil",
-        "//internal/httpcli",
         "//internal/jsonc",
         "//internal/lazyregexp",
-        "//internal/oauthutil",
         "//internal/own/codeowners/v1:codeowners",
         "//internal/own/types",
         "//internal/randstring",
@@ -169,7 +166,6 @@ go_library(
         "@io_opentelemetry_go_otel//attribute",
         "@org_golang_google_protobuf//proto",
         "@org_golang_x_crypto//bcrypt",
-        "@org_golang_x_oauth2//:oauth2",
         "@org_golang_x_sync//errgroup",
         "@org_golang_x_time//rate",
     ],
@@ -218,7 +214,6 @@ go_test(
         "main_test.go",
         "namespace_permissions_test.go",
         "namespaces_test.go",
-        "oauth_token_helper_test.go",
         "org_invitations_test.go",
         "org_members_db_test.go",
         "orgs_test.go",
@@ -291,7 +286,6 @@ go_test(
         "//internal/extsvc/gitlab",
         "//internal/featureflag",
         "//internal/jsonc",
-        "//internal/oauthutil",
         "//internal/own/codeowners/v1:codeowners",
         "//internal/own/types",
         "//internal/rbac/types",
