+++ internal/database/oauthtoken/oauth_token_helper.go
@@ -1,10 +1,11 @@
-package database
+package oauthtoken
 
 import (
 	"context"
 	"strconv"
 	"time"
 
+	"github.com/sourcegraph/sourcegraph/internal/database"
 	"github.com/sourcegraph/sourcegraph/internal/encryption"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc/auth"
 	"github.com/sourcegraph/sourcegraph/internal/extsvc/gitlab"
@@ -15,9 +16,9 @@ import (
 	"golang.org/x/oauth2"
 )
 
-// ExternalAccountTokenRefresher returns an oauthutil.TokenRefresher for the
+// externalAccountTokenRefresher returns an oauthutil.TokenRefresher for the
 // given external account.
-func ExternalAccountTokenRefresher(store UserExternalAccountsStore, externalAccountID int32, originalToken *auth.OAuthBearerToken) oauthutil.TokenRefresher {
+func externalAccountTokenRefresher(store database.UserExternalAccountsStore, externalAccountID int32, originalToken *auth.OAuthBearerToken) oauthutil.TokenRefresher {
 	return func(ctx context.Context, doer httpcli.Doer, oauthCtx oauthutil.OAuthContext) (token *auth.OAuthBearerToken, err error) {
 		// Start a transaction so that multiple refreshes don't happen simultaneously
 		tx, err := store.Transact(ctx)
@@ -74,9 +75,9 @@ func ExternalAccountTokenRefresher(store UserExternalAccountsStore, externalAcco
 	}
 }
 
-// ExternalServiceTokenRefresher returns an oauthutil.TokenRefresher for the
+// externalServiceTokenRefresher returns an oauthutil.TokenRefresher for the
 // given external service.
-func ExternalServiceTokenRefresher(db DB, externalServiceID int64, refreshToken string) oauthutil.TokenRefresher {
+func externalServiceTokenRefresher(db database.DB, externalServiceID int64, refreshToken string) oauthutil.TokenRefresher {
 	return func(ctx context.Context, doer httpcli.Doer, oauthCtx oauthutil.OAuthContext) (token *auth.OAuthBearerToken, err error) {
 		defer func() {
 			success := err == nil
@@ -126,9 +127,9 @@ func ExternalServiceTokenRefresher(db DB, externalServiceID int64, refreshToken 
 	}
 }
 
-func GetServiceRefreshAndStoreOAuthTokenFunc(db DB, externalServiceID int64, oauthContext *oauthutil.OAuthContext) func(context.Context, httpcli.Doer, *auth.OAuthBearerToken) (string, string, time.Time, error) {
+func GetServiceRefreshAndStoreOAuthTokenFunc(db database.DB, externalServiceID int64, oauthContext *oauthutil.OAuthContext) func(context.Context, httpcli.Doer, *auth.OAuthBearerToken) (string, string, time.Time, error) {
 	return func(ctx context.Context, cli httpcli.Doer, a *auth.OAuthBearerToken) (string, string, time.Time, error) {
-		tokenRefresher := ExternalServiceTokenRefresher(db, externalServiceID, a.RefreshToken)
+		tokenRefresher := externalServiceTokenRefresher(db, externalServiceID, a.RefreshToken)
 		token, err := tokenRefresher(ctx, cli, *oauthContext)
 		if err != nil {
 			return "", "", time.Time{}, err
@@ -138,9 +139,9 @@ func GetServiceRefreshAndStoreOAuthTokenFunc(db DB, externalServiceID int64, oau
 	}
 }
 
-func GetAccountRefreshAndStoreOAuthTokenFunc(store UserExternalAccountsStore, externalAccountID int32, oauthContext *oauthutil.OAuthContext) func(context.Context, httpcli.Doer, *auth.OAuthBearerToken) (string, string, time.Time, error) {
+func GetAccountRefreshAndStoreOAuthTokenFunc(store database.UserExternalAccountsStore, externalAccountID int32, oauthContext *oauthutil.OAuthContext) func(context.Context, httpcli.Doer, *auth.OAuthBearerToken) (string, string, time.Time, error) {
 	return func(ctx context.Context, cli httpcli.Doer, a *auth.OAuthBearerToken) (string, string, time.Time, error) {
-		tokenRefresher := ExternalAccountTokenRefresher(store, externalAccountID, a)
+		tokenRefresher := externalAccountTokenRefresher(store, externalAccountID, a)
 		token, err := tokenRefresher(ctx, cli, *oauthContext)
 		if err != nil {
 			return "", "", time.Time{}, err
